+++ lib/unicode/segment.ex
@@ -155,7 +155,7 @@ defmodule Unicode.String.Segment do
   end
 
   @doc """
-  Evaludates a list of rules against a given
+  Evaluates a list of rules against a given
   string.
 
   """